Remove api/v1/
