--- app/controllers/api/v1/auth/sessions_controller.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    module Auth
-      class SessionsController < ::DeviseTokenAuth::SessionsController
-        # Prevent session parameter from being passed
-        # Unpermitted parameter: session
-        wrap_parameters format: []
-      end
-    end
-  end
-end
