+++ config/routes.rb
@@ -1,12 +1,8 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  namespace :api do
-    namespace :v1 do
-      mount_devise_token_auth_for 'User', at: 'auth', controllers: {
-        registrations: 'api/v1/auth/registration',
-        sessions: 'api/v1/auth/sessions'
-      }
-    end
-  end
+  mount_devise_token_auth_for 'User', at: 'auth', controllers: {
+    registrations: 'auth/registration',
+    sessions: 'auth/sessions'
+  }
 end
